Add files via upload
The task is based on the paper from Osborn and Frey on the “Future of Employment”. The authors use a simple machine learning model and the ONET database to estimate a probability that each occupation will be automated. The ONET database is a survey of experts and current employees about the requirements of each occupation (job). Effectively, Osborn and Frey use the ONET data to predict an occupation’s probability of automation. Rather than predicting the probability of automation, my task will be to predict the probability of being ANALYTICAL.